update to work on macOS

+++ examples/api/run_query_1d_gtl.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ -z "$UCVM_INSTALL_PATH" ]; then
+  echo "Need to set UCVM_INSTALL_PATH to run >" ${0##*/} 
+  exit
+fi
+source $UCVM_INSTALL_PATH/conf/ucvm_env.sh
+
 expect=$(mktemp) || exit 1
 result=$(mktemp) || (trap 'rm -f "$expect"'; exit 1)
